requirements.txt
#kafka-python 
requests 
beautifulsoup4 
twilio 
python-dotenv
confluent_kafka